requirements.txt
alembic==1.9.1
aniso8601==9.0.1
blinker==1.5
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
loginpass==0.5
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wincertstore==0.2
WTForms==3.0.1
zipp==3.11.0
